Move the locale data context
