+++ src/impact-dashboard/EpidemicModelContext.tsx
@@ -2,7 +2,7 @@ import { sum } from "d3";
 import { pick } from "lodash";
 import React from "react";
 
-import { LocaleData } from "../locale-data-context";
+import { LocaleData } from "../contexts/locale-data-context";
 import { Facility } from "../page-multi-facility/types";
 
 export type PlannedRelease = { date?: Date; count?: number };
