+++ src/providers/site-provider/SiteProvider.tsx
@@ -4,10 +4,10 @@ import { ToastProvider } from "react-toast-notifications";
 
 import AppAuth0ClientPromise from "../../auth/AppAuth0ClientPromise";
 import { Auth0Provider } from "../../auth/react-auth0-spa";
+import { LocaleDataProvider } from "../../contexts/locale-data-context";
 import { ScenarioProvider } from "../../contexts/scenario-context";
 import Toast from "../../design-system/Toast";
 import { FeatureFlagsProvider } from "../../feature-flags";
-import { LocaleDataProvider } from "../../locale-data-context";
 import {
   FacilityContext,
   rtDataReducer,
